+++ src/django/api/management/commands/post_deployment.py
@@ -11,3 +11,4 @@ class Command(BaseCommand):
         call_command('migrate')
         call_command('reindex_database')
         call_command('reindex_locations_with_approved_claim')
+        call_command('backfill_moderation_event_os_id')